+++ ios/example/app.js
@@ -113,7 +113,7 @@ function requestTrackingAuthorization() {
 function requestConsent() {
   console.log("request consent");
   Admob.requestConsentInfoUpdateWithParameters({
-    testDeviceIdentifiers: [Admob.SIMULATOR_ID, 'YOUR-TEST-DEVICE-HASHED-ID'],
+    testDeviceIdentifiers: ['YOUR-TEST-DEVICE-HASHED-ID'],
     geography: Admob.DEBUG_GEOGRAPHY_EEA, // `Admob.DEBUG_GEOGRAPHY_NOT_EEA` or `Admob.DEBUG_GEOGRAPHY_DISABLED`
     tagForUnderAgeOfConsent: false,
     callback: function (e) {
@@ -146,7 +146,12 @@ function requestConsent() {
             }
             // If the status is "obtained" (freshly granted) or not required (already granted) continue
             if ([Admob.CONSENT_STATUS_NOT_REQUIRED, Admob.CONSENT_STATUS_OBTAINED].includes(e.status)) {
-              openTestAdsWin();
+              if (Admob.canShowAds()){
+                openTestAdsWin();
+              } else {
+                alert('You have not granted at least the minimum requirements to show ads!' +
+                  'No fear! You can buy an in-app purchase to use the app without ads :)');
+              }                
             } else {
               alert('Not ready to show ads! Status = ' + e.status);
             }
@@ -250,10 +255,13 @@ function openTestAdsWin() {
     contentURL: 'https://admob.com', // URL string for a webpage whose content matches the app content.
     requestAgent: 'Titanium Mobile App', // String that identifies the ad request's origin.
     extras: {
+      'npa': !Admob.canShowPersonalizedAds(),
       'version': 1.0,
       'name': 'My App'
     }, // Object of additional infos
-    tagForChildDirectedTreatment: false, // http:///business.ftc.gov/privacy-and-security/childrens-privacy for more infos
+    tagForChildDirectedTreatment: false, // https://developers.google.com/admob/ios/targeting#child-directed_setting for more infos
+    tagForUnderAgeOfConsent: false, //https://developers.google.com/admob/ios/targeting#users_under_the_age_of_consent for more infos
+    maxAdContentRating: Admob.MAX_AD_CONTENT_RATING_GENERAL, // https://developers.google.com/admob/ios/targeting#ad_content_filtering for more infos
     keywords: ['keyword1', 'keyword2']
   });
 
@@ -289,10 +297,14 @@ function openTestAdsWin() {
     adUnitId: 'ca-app-pub-3940256099942544/4411468910', // You can get your own at http: //www.admob.com/
     keywords: ['keyword1', 'keyword2'],
     extras: {
+      'npa': !Admob.canShowPersonalizedAds(),
       'version': 1.0,
       'name': 'My App'
     }, // Object of additional infos
-    visible: false // If true, covers the win when added and can't tap nothing
+    visible: false, // If true, covers the win when added and can't tap nothing
+    tagForChildDirectedTreatment: false, // https://developers.google.com/admob/ios/targeting#child-directed_setting for more infos
+    tagForUnderAgeOfConsent: false, //https://developers.google.com/admob/ios/targeting#users_under_the_age_of_consent for more infos
+    maxAdContentRating: Admob.MAX_AD_CONTENT_RATING_GENERAL, // https://developers.google.com/admob/ios/targeting#ad_content_filtering for more infos
   });
   
 
@@ -352,9 +364,13 @@ function openTestAdsWin() {
     adType: Admob.AD_TYPE_REWARDED_VIDEO,
     adUnitId: 'ca-app-pub-3940256099942544/1712485313',
     extras: {
+      'npa': !Admob.canShowPersonalizedAds(),
       'version': 1.0,
       'name': 'My App'
-    } // Object of additional infos
+    }, // Object of additional infos
+    tagForChildDirectedTreatment: false, // https://developers.google.com/admob/ios/targeting#child-directed_setting for more infos
+    tagForUnderAgeOfConsent: false, //https://developers.google.com/admob/ios/targeting#users_under_the_age_of_consent for more infos
+    maxAdContentRating: Admob.MAX_AD_CONTENT_RATING_GENERAL, // https://developers.google.com/admob/ios/targeting#ad_content_filtering for more infos
   });
  
   rewardedVideo.addEventListener('didRewardUser', function (reward) {
@@ -442,9 +458,13 @@ function openTestAdsWin() {
       adType: Admob.AD_TYPE_APP_OPEN,
       adUnitId: 'ca-app-pub-3940256099942544/5662855259', // You can get your own at http: //www.admob.com/
       extras: {
+        'npa': !Admob.canShowPersonalizedAds(),
         'version': 1.0,
         'name': 'My App'
-      } // Object of additional infos
+      }, // Object of additional infos
+      tagForChildDirectedTreatment: false, // https://developers.google.com/admob/ios/targeting#child-directed_setting for more infos
+      tagForUnderAgeOfConsent: false, //https://developers.google.com/admob/ios/targeting#users_under_the_age_of_consent for more infos
+      maxAdContentRating: Admob.MAX_AD_CONTENT_RATING_GENERAL, // https://developers.google.com/admob/ios/targeting#ad_content_filtering for more infos
     });
 
     // appOpenAd custom events